api/pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ukde-api"
version = "0.1.0"
description = "FastAPI bootstrap for UKDE."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.135.1",
  "httpx==0.28.1",
  "pillow==11.3.0",
  "python-multipart==0.0.20",
  "psycopg[binary]==3.2.13",
  "pydantic-settings==2.13.1",
  "uvicorn==0.41.0"
]

[project.optional-dependencies]
dev = [
  "pytest==9.0.2"
]

[project.scripts]
ukde-bundle-verify = "app.exports.bundle_verify_cli:run_cli"
ukde-provenance-replay = "app.exports.replay_cli:run_cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
